refactor: busca com valor pre definido
